Add files via upload
Virtual Key for Repository is a general purpose prototyped application provides user to perform multiple actions on the repository: List files in directory, Business level operations like: Add file in directory, Delete file from directory, Search for a file in directory etc. It handles all the exception (both system-defined and user-defined) that comes into the way. Algorithms and techniques are used to increase efficiency of the application. An application continues to run until user don't close the application.